add the status code

+++ routes/route.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const router = express.Router()
 const Ecom = require('../models/model')
-
+const mongoose = require("mongoose");
 
 router.get('/', async(req,res) => {
     try{
@@ -14,9 +14,14 @@ router.get('/', async(req,res) => {
 })
 
 router.get('/:id', async(req,res) => {
-    try{
-           const ecom = await Ecom.findById(req.params.id)
-           res.status(200).json(ecom)
+    try{   
+        if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+            return res.status(400).send("Invalid Item id");
+        } else{
+            const ecom = await Ecom.findById(req.params.id)
+            res.status(200).json(ecom)
+        }
+           
     }catch(err){
         res.send('Error ' + err)
     }
@@ -41,11 +46,16 @@ router.post('/', async(req,res) => {
 
 router.put('/:id',async(req,res)=> {
     try{
-        const ecom = await Ecom.findById(req.params.id) 
-        ecom.price = req.body.price
-        ecom.quantity = req.body.quantity
-        const x = await ecom.save()
-        res.status(200).json(x)   
+        if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+            return res.status(400).send("Invalid Item id");
+        } else{
+            const ecom = await Ecom.findById(req.params.id) 
+            ecom.price = req.body.price
+            ecom.quantity = req.body.quantity
+            const x = await ecom.save()
+            res.status(200).json(x)
+        }
+           
     }catch(err){
         res.send('Error')
     }
@@ -54,8 +64,13 @@ router.put('/:id',async(req,res)=> {
 
 router.delete('/:id',async(req,res)=> {
     try{
-        const ecom = await Ecom.findByIdAndRemove(req.params.id)
-        res.status(204).json(ecom)
+        if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+            return res.status(400).send("Invalid Item id");
+        } else{
+            const ecom = await Ecom.findByIdAndRemove(req.params.id)
+            res.status(204).json(ecom)
+        }
+        
     }catch(err){
         res.send('Error')
     }
